pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
alembic = "^1.6.2"
fastapi = "^0.64.0"
uvicorn = "^0.13.4"
httpx = "^0.18.1"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.15.1"
loguru = "^0.5.3"
SQLAlchemy = "^1.4.15"
asyncpg = "^0.23.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
